zephyr: device: Split gpio and flash to own files
Move this code out of the device.rs file, and into separate files for
each module.
